+++ assets/javascripts/kitten/components/form/field/components/radio-set.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { RadioSet } from '../../../form/radio-button-set'
+import { RadioSet } from '../../../form/radio-set'
 
 export const FieldRadioSet = props => (
   <div className="k-u-margin-top-single">